変更 - routerのindexに直接コーディングしていたので、内容をindexから個別に移動
