+++ src/features/Archives/router/index.tsx
@@ -1,57 +1,3 @@
-import React, { Suspense } from 'react';
-
-import { useLocation, useNavigate, useParams } from 'react-router-dom';
-
 import { Archive } from './Archive';
 
-import NextLoad from '../component/NextLoad';
-import { Hover } from '../../../component/Element/Button';
-import { LeftArrow } from '../../../component/Element/Icon';
-import { VtuberDetail } from '../types';
-import { Channel } from '../component/Channel';
-
-const typeGuard = (target: string | undefined) => {
-    if (target === undefined) {
-        return 'error';
-    }
-    return target;
-};
-interface LocationState {
-    channel: VtuberDetail;
-}
-
-export const ArchiveRouter = () => {
-    const { channelID } = useParams<'channelID'>();
-    const location = useLocation();
-    const locationState = location.state as LocationState;
-    const navigate = useNavigate();
-
-    //TypeGuardでundefinedを除外する
-    const targetChannelID = typeGuard(channelID);
-
-    return (
-        <div className="flex flex-col">
-            <div className="flex justify-start">
-                <Hover handler={() => navigate(-1)} size="small" radius="full">
-                    <div className="text-gray-100">
-                        <LeftArrow />
-                    </div>
-                </Hover>
-            </div>
-            <div className="grid justify-items-center">
-                <Channel channelDetail={locationState.channel}></Channel>
-            </div>
-            <Suspense fallback={<div></div>}>
-                <Archive
-                    name={locationState.channel.name}
-                    channelId={targetChannelID}
-                />
-            </Suspense>
-            <div className="mt-2 mb-2">
-                <Suspense fallback={<p>Loading...</p>}>
-                    <NextLoad channelId={targetChannelID} />
-                </Suspense>
-            </div>
-        </div>
-    );
-};
+export { Archive };
